feat: split iOS app to nutrition-only (#513) — Phase 1
Removed all workout UI from the iOS app:
- Deleted 15 workout files (ActiveWorkout, Plans, Dashboard, Progress,
  workout Components)
- Rebuilt MainTabView: 3 tabs (Nutrition, Activity, Settings)
- New ActivityView: read-only workout list + body weight chart
- Stripped UIModels.swift of workout types (UISet, UIExercise, plate math, etc.)
- Stripped WorkoutModels.swift of plan/progress/template types
- Stripped SettingsView of workout settings (rest timers, bar weights,
  plate math, deload, progression)
- Updated xcodeproj to match

iOS is now a nutrition-focused app that reads workout data from the
shared backend (read-only).
